Create constructor for DTO
